Handle holes in arguments
Patched an issue in `Array.concat(Deep)`, `Dictionary.merge(Deep)` and `Set.merge` where they would stop at a `nil` value in the arguments.

This is because, when using `ipairs`, it automatically assumes the end of the array has been reached when it hits a `nil` value.

The new implementation uses `select` to process the vararg (`...`) in order to continue processing even if a `nil` value is present.